re-organize CLI
