--- src/commands/index.ts
@@ -1,4 +1,0 @@
-import * as init from './init'
-import * as packages from './packages'
-
-export const commands = [init, packages]
