--- src/commands/packages.ts
@@ -1,8 +1,0 @@
-export const command = 'packages [packageNames...]'
-export const desc = 'Specify list of packages you want to link'
-export const builder = {
-  packageNames: {
-    array: true,
-  },
-}
-export function handler() {}
